moved class into appropriate file
